+++ src/components/ApplicationViews.js
@@ -7,6 +7,7 @@ import { UserPostsList } from "./posts/UserPosts"
 import { TagList } from "./tags/TagList"
 import { CategoriesList } from "./categories/CategoriesList"
 import { UserList } from "./users/UserList"
+import { EditPosts } from "./posts/EditPost"
 
 
 export const ApplicationViews = () => {
@@ -32,6 +33,9 @@ export const ApplicationViews = () => {
     </Route>
     <Route exact path="/users">
       <UserList />
+    </Route>
+    <Route exact path="/posts/:postId(\d+)/edit">
+      <EditPosts />
     </Route>
     </>
   )
