+++ src/components/posts/UserPosts.js
@@ -1,25 +1,19 @@
 import React, { useEffect, useState } from "react";
-import { getUserPost, deletePost, editPost } from "./PostManager";
-import { getCategories } from "../categories/CategoriesManager";
+import { getUserPost, deletePost } from "./PostManager";
+
+import { Link } from "react-router-dom";
 
 
 export const UserPostsList = () => {
     const [posts, setPosts] = useState([])
     const [currentUser, setUser] = useState()
     const [refresh, setRefresh] = useState(false)
-    const [editable, setEditable] = useState(false)
-    const [newPost, setNewPost] = useState({})
-    const [categories, setCategories] = useState([])
 
 
     useEffect(() => {
         setUser(localStorage.getItem("token"))
     }, [refresh])
 
-    useEffect(() => {
-        getCategories()
-            .then((data) => setCategories(data) )
-    }, [])
 
     useEffect(() => {
         getUserPost(currentUser)
@@ -27,16 +21,6 @@ export const UserPostsList = () => {
             .then(() => setRefresh(false))
     }, [currentUser, refresh])
 
-    const submitPost = (post) => {
-        const copy = {...newPost}
-        copy.userId = post.userId
-        copy.publicationDate = post.publicationDate
-        copy.approved = post.approved
-        setNewPost(copy)
-        editPost(newPost, post.id)
-        .then(setEditable(false))
-        .then(setRefresh(true))
-    }
 
     return (
         <>
@@ -45,75 +29,31 @@ export const UserPostsList = () => {
                 {posts.map(
                     (post) => {
                         return <>
-                            {editable === false ?
-                                <li className="card post--list" key={`post--${post.id}`}>
-                                    <div key={`post--${post.id}`}>
-                                        <h2 className="post--title">
-                                            {post.title}
-                                        </h2>
-                                        <div className="post--user">
-                                            by {post.user?.fullName}
-                                        </div>
-                                        <div className="post--category">
-                                            Category: {post.category.label}
-                                        </div>
-                                        <div className="post--date">
-                                            {post.publicationDate}
-                                        </div>
-                                        <img className="post--image" src={post.imageURL} alt={post.title}
-                                        />
-                                        <div className="post--content">
-                                            {post.content}
-                                        </div>
-                                        <button onClick={() =>{
-                                            setNewPost(post)
-                                            setEditable(true)}}>Edit</button>
-                                        <button onClick={() => {
-                                            deletePost(post.id, setRefresh)
-                                        }}>Delete</button>
-                                
-                                </div>
-                                </li>
-                                
-                                : <li className="card post--list" key={`post--${post.id}`}>
+                            <li className="card post--list" key={`post--${post.id}`}>
                                 <div key={`post--${post.id}`}>
-                                    <input type ="text" defaultValue={post.title} className="post--title" onChange={e => {
-                                        const copy = {...newPost}
-                                        copy.title = e.target.value
-                                        setNewPost(copy)
-                                    }}/>
+                                    <h2 className="post--title">
+                                        {post.title}
+                                    </h2>
                                     <div className="post--user">
                                         by {post.user?.fullName}
                                     </div>
-                                    <select className="post--category" defaultValue={post.categoryId} onChange={e => {
-                                        const copy = {...newPost}
-                                        copy.categoryId = parseInt(e.target.value)
-                                        setNewPost(copy)
-                                    }}>
-                                        {categories.map(category => {
-                                            return <option key={`category--${category.id}`} value={category.id}>{category.label}</option>
-                                        })}
-                                    </select>
+                                    <div className="post--category">
+                                        Category: {post.category.label}
+                                    </div>
                                     <div className="post--date">
                                         {post.publicationDate}
                                     </div>
-                                    <textarea className="post--image" defaultValue={post.imageURL} onChange={e => {
-                                        const copy = {...newPost}
-                                        copy.imageURL = e.target.value
-                                        setNewPost(copy)
-                                    }}
-                                    /><br></br>
-                                    <textarea className="post--content" defaultValue={post.content} onChange={e => {
-                                        const copy = {...newPost}
-                                        copy.content = e.target.value
-                                        setNewPost(copy)
-                                    }}
-                                    /><br></br>
-                                    <button onClick={e => {
-                                       submitPost(post)
-                                    }}>Submit Changes</button>
+                                    <img className="post--image" src={post.imageURL} alt={post.title}
+                                    />
+                                    <div className="post--content">
+                                        {post.content}
+                                    </div>
+                                    <button><Link to={`/posts/${post.id}/edit`}>Edit</Link></button>
+                                    <button onClick={() => {
+                                        deletePost(post.id, setRefresh)
+                                    }}>Delete</button>
                                 </div>
-                            </li>}
+                            </li>
                         </>
                     }
                 )}
